ingest/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "egograph-ingest"
version = "0.1.0"
description = "Data ingestion service for EgoGraph"
requires-python = ">=3.13"
dependencies = [
    "egograph",
    "spotipy>=2.25.0",
    "llama-index>=0.12.0",
    "llama-index-core>=0.12.0",
    "qdrant-client>=1.12.0",
    "requests>=2.32.0",
    "python-dateutil>=2.9.0",
    "tenacity>=9.0.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",
    "python-dotenv>=1.0.0",
]

[tool.hatch.metadata]
allow-direct-references = true

[project.scripts]
egograph-ingest = "src.main:main"

[tool.uv.sources]
egograph = { workspace = true }

[dependency-groups]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.25.0",
    "pytest-cov>=6.0.0",
    "black>=24.10.0",
    "flake8>=7.1.0",
    "mypy>=1.14.0",
    "responses>=0.25.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
addopts = "-v --cov=src"
